+++ pkg/provider/surge.go
@@ -46,8 +46,6 @@ func checkSurgeSupport(p proxy.Proxy) bool {
 		}
 	case *proxy.Trojan:
 		return true
-	case *proxy.Vless:
-		return true
 	default:
 		return false
 	}